REFACTOR: 멀티 모듈 적용
