+++ batch/src/main/java/jshop/batch/BatchApplication.java
@@ -1,13 +1,13 @@
-package jshop.web;
+package jshop.batch;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class WebApplication {
+public class BatchApplication {
 
     public static void main(String[] args) {
-        SpringApplication.run(WebApplication.class, args);
+        SpringApplication.run(BatchApplication.class, args);
     }
 
 }
