+++ batch/src/test/java/jshop/batch/BatchApplicationTests.java
@@ -1,0 +1,13 @@
+package jshop.batch;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BatchApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
