--- common/src/test/java/jshop/common/CommonApplicationTests.java
@@ -1,13 +1,0 @@
-package jshop.common;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CommonApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
