+++ core/src/test/resources/application.yml
@@ -1,0 +1,4 @@
+spring:
+  jpa:
+    hibernate:
+      ddl-auto: update
