--- infra/src/test/java/jshop/infra/InfraApplicationTests.java
@@ -1,13 +1,0 @@
-package jshop.infra;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class InfraApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
